SHDP-525 Move all namespace classes to new o.s.d.h.config.namespace package
