+++ spring-hadoop-config/src/main/java/org/springframework/data/hadoop/config/namespace/HadoopConfigParser.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2013 the original author or authors.
+ * Copyright 2011 the original author or authors.
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,29 +13,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.data.hadoop.config;
+package org.springframework.data.hadoop.config.namespace;
 
 import org.springframework.beans.factory.support.BeanDefinitionBuilder;
 import org.springframework.beans.factory.xml.ParserContext;
+import org.springframework.data.hadoop.configuration.ConfigurationFactoryBean;
 import org.w3c.dom.Element;
 
 /**
+ * Simple namespace parser for hadoop:config.
+ * 
  * @author Costin Leau
  */
-public class AbstractGenericOptionsParser extends AbstractPropertiesConfiguredBeanDefinitionParser {
+class HadoopConfigParser extends AbstractPropertiesConfiguredBeanDefinitionParser {
 
+	@Override
+	protected Class<?> getBeanClass(Element element) {
+		return ConfigurationFactoryBean.class;
+	}
+
+	@Override
 	protected boolean isEligibleAttribute(String attributeName) {
-		return !("files".equals(attributeName) || "libs".equals(attributeName) || "archives".equals(attributeName))
-				&& super.isEligibleAttribute(attributeName);
+		return (!"resources".equals(attributeName)) && super.isEligibleAttribute(attributeName);
 	}
 
 	@Override
 	protected void doParse(Element element, ParserContext parserContext, BeanDefinitionBuilder builder) {
-		// parse attributes using conventions
 		super.doParse(element, parserContext, builder);
 
-		NamespaceUtils.setCSVProperty(element, builder, "files");
-		NamespaceUtils.setCSVProperty(element, builder, "libs");
-		NamespaceUtils.setCSVProperty(element, builder, "archives");
+		NamespaceUtils.setCSVProperty(element, builder, "resources");
+	}
+
+	@Override
+	protected String defaultId(ParserContext context, Element element) {
+		return "hadoopConfiguration";
 	}
 }
