+++ spring-hadoop-config/src/main/java/org/springframework/data/hadoop/config/namespace/HadoopStreamJobParser.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2011 the original author or authors.
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.springframework.data.hadoop.config.namespace;
+
+import org.springframework.beans.factory.support.BeanDefinitionBuilder;
+import org.springframework.beans.factory.xml.ParserContext;
+import org.springframework.data.hadoop.mapreduce.StreamJobFactoryBean;
+import org.springframework.util.StringUtils;
+import org.springframework.util.xml.DomUtils;
+import org.w3c.dom.Element;
+
+/**
+ * Job declaration parser.
+ * 
+ * @author Costin Leau
+ */
+class HadoopStreamJobParser extends AbstractGenericOptionsParser {
+
+	@Override
+	protected Class<?> getBeanClass(Element element) {
+		return StreamJobFactoryBean.class;
+	}
+
+	@Override
+	protected void doParse(Element element, ParserContext parserContext, BeanDefinitionBuilder builder) {
+		// parse attributes using conventions
+		super.doParse(element, parserContext, builder);
+
+		// parse properties
+		String props = DomUtils.getChildElementValueByTagName(element, "cmd-env");
+
+		if (StringUtils.hasText(props)) {
+			builder.addPropertyValue("cmdEnv", props);
+		}
+	}
+}
