+++ spring-hadoop-config/src/main/java/org/springframework/data/hadoop/config/namespace/HiveRunnerParser.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.data.hadoop.config;
+package org.springframework.data.hadoop.config.namespace;
 
 import java.util.Collection;
 import java.util.Collections;
