+++ spring-hadoop-config/src/main/java/org/springframework/data/hadoop/config/namespace/PigTaskletParser.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011 the original author or authors.
+ * Copyright 2011-2013 the original author or authors.
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,42 +13,42 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.data.hadoop.config;
+package org.springframework.data.hadoop.config.namespace;
 
 import java.util.Collection;
 
+import org.springframework.beans.factory.config.BeanDefinition;
 import org.springframework.beans.factory.support.BeanDefinitionBuilder;
 import org.springframework.beans.factory.xml.ParserContext;
-import org.springframework.data.hadoop.hive.HiveClientFactoryBean;
+import org.springframework.data.hadoop.batch.pig.PigTasklet;
 import org.springframework.util.CollectionUtils;
 import org.w3c.dom.Element;
 
 /**
- * Parser for "hive-client" element.
+ * Parser for 'pig-tasklet' element.
  * 
  * @author Costin Leau
  */
-class HiveClientParser extends AbstractImprovedSimpleBeanDefinitionParser {
+class PigTaskletParser extends AbstractImprovedSimpleBeanDefinitionParser {
 
 	@Override
 	protected Class<?> getBeanClass(Element element) {
-		return HiveClientFactoryBean.class;
-	}
-
-	@Override
-	protected String defaultId(ParserContext context, Element element) {
-		return "hiveClientFactory";
+		return PigTasklet.class;
 	}
 
 	@Override
 	protected void doParse(Element element, ParserContext parserContext, BeanDefinitionBuilder builder) {
 		// parse attributes using conventions
 		super.doParse(element, parserContext, builder);
-
 		// parse scripts
-		Collection<Object> scripts = HiveRunnerParser.parseScripts(parserContext, element);
+		Collection<BeanDefinition> scripts = PigServerParser.parseScripts(parserContext, element);
 		if (!CollectionUtils.isEmpty(scripts)) {
 			builder.addPropertyValue("scripts", scripts);
 		}
+	}
+
+	@Override
+	protected boolean shouldGenerateIdAsFallback() {
+		return true;
 	}
 }
