+++ spring-hadoop-config/src/main/resources/META-INF/spring.handlers
@@ -1,1 +1,1 @@
-http\://www.springframework.org/schema/hadoop=org.springframework.data.hadoop.config.HadoopNamespaceHandler
+http\://www.springframework.org/schema/hadoop=org.springframework.data.hadoop.config.namespace.HadoopNamespaceHandler
